requirements.txt
bcrypt==3.1.3
cffi==1.10.0
Django==1.10
pbr==3.0.1
pycparser==2.17
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
stevedore==1.23.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
